Extract navigation constants to a separate file
