+++ packages/react-ui/src/app/lib/authentication-session.ts
@@ -4,7 +4,7 @@ import { jwtDecode } from 'jwt-decode';
 import { authenticationApi } from '@/app/lib/authentication-api';
 import { AuthenticationResponse, isNil } from '@openops/shared';
 import { NavigateFunction } from 'react-router-dom';
-import { LOGOUT_EVENT_KEY } from './navigation-events';
+import { LOGOUT_EVENT_KEY } from './navigation-constants';
 import { navigationUtil } from './navigation-util';
 import { projectAuth } from './project-auth';
