+++ packages/react-ui/src/app/lib/navigation-constants.ts
@@ -1,0 +1,3 @@
+export const LAST_VISITED_KEY = 'last_visited_page';
+export const LOGOUT_EVENT_KEY = 'logout_event';
+export const LOGOUT_FLAG_KEY = 'logout_flag';
